Style, documentation, angular trees

+++ src/com/tagbio/umap/metric/EuclideanMetric.java
@@ -19,9 +19,9 @@ public class EuclideanMetric extends Metric {
   @Override
   public double distance(final float[] x, final float[] y) {
     //  D(x, y) = \sqrt{\sum_i (x_i - y_i)^2}
-    double result = 0.0;
+    float result = 0;
     for (int i = 0; i < x.length; ++i) {
-      final double d = x[i] - y[i];
+      final float d = x[i] - y[i];
       result += d * d;
     }
     return Math.sqrt(result);
